Fixing tests and cleaning up code
