+++ ai_diffusion/text.py
@@ -222,7 +222,7 @@ def edit_attention(text: str, positive: bool) -> str:
 
 
 # creates the img text metadata for embedding in PNG files in style of Automatic1111
-def create_img_metadata(params: "JobParams"):  # noqa F821  to avoid circular import
+def create_img_metadata(params: "JobParams"):  # noqa: F821  # type: ignore[name-defined]  to avoid circular import
     meta = params.metadata
 
     prompt = meta.get("prompt", "")
